backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain==0.3.0
langchain-community==0.3.0
langchain-experimental==0.3.0
langchain-deepseek==0.1.0
llama-index==0.11.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
tiktoken==0.7.0
beautifulsoup4==4.12.3
html2text==2024.2.26
requests==2.32.3
python-multipart==0.0.12
